Refactor clientAction logic in BranchConfig form by moving the clientAction from form.tsx to route.tsx and update the form.tsx to only contain the schema and form components.
